% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_dad_lrwgs}
\alias{add_dad_lrwgs}
\title{add_dad_lrwgs}
\usage{
add_dad_lrwgs(my_df, full_chdgenes)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}

\item{full_chdgenes}{A data.frame containing the full CHDGENES sample inventory}
}
\value{
Additional columns as denoted by the function title.
}
\description{
Adds a column to a data.frame denoting if a paternal sample for a proband blinded ID has been shipped for long read whole genome sequencing
}
