% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_dad_total_dna}
\alias{add_dad_total_dna}
\title{add_dad_total_dna}
\usage{
add_dad_total_dna(my_df, full_chdgenes)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}

\item{full_chdgenes}{A data.frame containing the full CHDGENES sample inventory}
}
\value{
Additional column as denoted by the function title. Note this is only QC pass DNA that is listed as either in circulation or in reserve.
}
\description{
add_dad_total_dna
}
