% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_family_id}
\alias{add_family_id}
\title{add_family_id}
\usage{
add_family_id(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}
}
\value{
Additional column with the 5 digit family ID
}
\description{
Adds the five digit family ID to an existing data.frame
}
