% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_in_trio}
\alias{add_in_trio}
\title{add_in_trio}
\usage{
add_in_trio(my_df, full_chdgenes)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}

\item{full_chdgenes}{A data.frame containing the full CHDGENES sample inventory}
}
\value{
Additional column denoting if the blinded ID is contained in a biobanking trio (e.g., at least one specimen was submitted for the proband and both parents).
}
\description{
Adds a column to a data.frame denoting if the provided blinded IDs are part of a trio.
}
