% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_shipped_lrwgs}
\alias{add_shipped_lrwgs}
\title{add_shipped_lrwgs}
\usage{
add_shipped_lrwgs(my_df, full_chdgenes)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}

\item{full_chdgenes}{A data.frame containing the full CHDGENES sample inventory}
}
\value{
Additional columns denoting if the blinded ID has been shipped for lrWGS, the date of shipment, and the external order ID number.
}
\description{
add_shipped_lrwgs
}
