% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_specimens_in_biobank}
\alias{add_specimens_in_biobank}
\title{add_specimens_in_biobank}
\usage{
add_specimens_in_biobank(my_df, full_chdgenes)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}

\item{full_chdgenes}{A data.frame containing the full CHDGENES sample inventory}
}
\value{
Additional column as denoting if at least one specimen was received by the biobank for the provided blinded IDs.
}
\description{
add_specimens_in_biobank
}
