% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_total_avail_dna}
\alias{add_total_avail_dna}
\title{add_total_avail_dna}
\usage{
add_total_avail_dna(my_df, full_chdgenes)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}

\item{full_chdgenes}{A data.frame containing the full CHDGENES sample inventory}
}
\value{
Additional column as denoted by the function title. This will add columns for all blind ID (proband or parent) passed to the function. Note this is only QC pass DNA that is listed as either in circulation or in reserve.
}
\description{
add_total_avail_dna
}
