% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_in_trio}
\alias{get_in_trio}
\title{get_in_trio}
\usage{
get_in_trio(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the a list of blind IDs. Often this will be the full CHDGENES sample inventory (e.g., aliquot level data).}
}
\value{
A data.frame with columns as denoted by the function title.
}
\description{
Generates a data.frame of family IDs and a column denoting whether the ID is part of a complete trio (proband, mom, dad)
}
