% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_probands_only}
\alias{get_probands_only}
\title{get_probands_only}
\usage{
get_probands_only(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the a list of blind IDs. Often this will be the full CHDGENES sample inventory (e.g., aliquot level data).}
}
\value{
A data.frame with columns as denoted by the function title.
}
\description{
Removes non-proband blinded IDs from a data.frame
}
