% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_shipped_wes}
\alias{get_shipped_wes}
\title{get_shipped_wes}
\usage{
get_shipped_wes(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the full CHDGENES sample inventory (e.g., aliquot level data).}
}
\value{
A data.frame with columns as denoted by the function title.
}
\description{
Generates a data.frame of blinded IDs shipped for exome sequencing
}
