% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_source_type}
\alias{get_source_type}
\title{get_source_type}
\usage{
get_source_type(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the full CHDGENES sample inventory (e.g., aliquot level data).}
}
\value{
A data.frame with columns as denoted by the function title.
}
\description{
Generates a data.frame of blinded IDs and the source type of the submitted specimens. If a blinded ID has multiple submission types they will be concatenated (e.g., wb|sal).
}
