% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_total_avail_dna}
\alias{get_total_avail_dna}
\title{get_total_avail_dna}
\usage{
get_total_avail_dna(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the full CHDGENES sample inventory (e.g., aliquot level data).}
}
\value{
A data.frame with columns as denoted by the function title.
}
\description{
get_total_avail_dna
}
