% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funs.R
\name{read_chdgenes}
\alias{read_chdgenes}
\title{read_chdgenes}
\usage{
read_chdgenes(my_path)
}
\arguments{
\item{my_path}{The full CHDGENES sample data inventory file as a semicolon delimited text file.}
}
\value{
The full CHDGENES sample data inventory file as a R data.frame
}
\description{
read_chdgenes
}
