% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funs.R
\name{remove_decommissioned}
\alias{remove_decommissioned}
\title{remove_decommissioned}
\usage{
remove_decommissioned(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the a list of blind IDs. Often this will be the full CHDGENES sample inventory (e.g., aliquot level data).}
}
\value{
A data.frame with all decommissioned samples removed.
}
\description{
remove_decommissioned
}
