% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funs.R
\name{remove_probands}
\alias{remove_probands}
\title{remove_probands}
\usage{
remove_probands(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the a list of blind IDs. Often this will be the full CHDGENES sample inventory (e.g., aliquot level data).}
}
\value{
A data.frame with all proband data removed.
}
\description{
remove_probands
}
