% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_funs.R
\name{remove_relatives}
\alias{remove_relatives}
\title{remove_relatives}
\usage{
remove_relatives(my_df)
}
\arguments{
\item{my_df}{A data.frame containing the a list of blind IDs. Often this will be the full CHDGENES sample inventory (e.g., aliquot level data).}
}
\value{
A data.frame with specimens from only probands, mom (-01), and dad (-02). All relative samples will be removed.
}
\description{
remove_relatives
}
