% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_mom_array}
\alias{add_mom_array}
\title{add_mom_array}
\usage{
add_mom_array(my_df, full_chdgenes)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}

\item{full_chdgenes}{A data.frame containing the full CHDGENES sample inventory}
}
\value{
Additional columns as denoted by the function title.
}
\description{
add_mom_array
}
