% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_mom_total_dna}
\alias{add_mom_total_dna}
\title{add_mom_total_dna}
\usage{
add_mom_total_dna(my_df, full_chdgenes)
}
\arguments{
\item{my_df}{A data.frame containing the blinded IDs of interest}

\item{full_chdgenes}{A data.frame containing the full CHDGENES sample inventory}
}
\value{
Additional column as denoted by the function title. Note this is only QC pass DNA that is listed as either in circulation or in reserve.
}
\description{
Adds a column to a data.frame denoting the total available QC pass DNA for the mother of a proband blinded ID
}
